changing over from coloring dir hash to checking coloring config vs. current of and wrt vars and sizes
